Adds support for setting up gRPC routes in tests. Updates gRPC tests to implement @SetupRoute methods.
