+++ webclient/tests/grpc/src/test/java/io/helidon/webclient/grpc/tests/GrpcTest.java
@@ -23,19 +23,18 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import com.google.protobuf.StringValue;
-import io.grpc.stub.StreamObserver;
+
 import io.helidon.common.configurable.Resource;
 import io.helidon.common.tls.Tls;
 import io.helidon.webclient.grpc.GrpcClient;
 import io.helidon.webclient.grpc.GrpcClientMethodDescriptor;
 import io.helidon.webclient.grpc.GrpcServiceDescriptor;
 import io.helidon.webserver.WebServer;
-import io.helidon.webserver.WebServerConfig;
-import io.helidon.webserver.grpc.GrpcRouting;
 import io.helidon.webserver.testing.junit5.ServerTest;
-import io.helidon.webserver.testing.junit5.SetUpServer;
+
 import org.junit.jupiter.api.Test;
+import com.google.protobuf.StringValue;
+import io.grpc.stub.StreamObserver;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -86,36 +85,6 @@ class GrpcTest extends GrpcBaseTest {
                                 .responseType(Strings.StringMessage.class)
                                 .build())
                 .build();
-    }
-
-    @SetUpServer
-    public static void setup(WebServerConfig.Builder builder) {
-        builder.tls(tls -> tls.privateKey(key -> key
-                                .keystore(store -> store
-                                        .passphrase("password")
-                                        .keystore(Resource.create("server.p12"))))
-                        .privateKeyCertChain(key -> key
-                                .keystore(store -> store
-                                        .trustStore(true)
-                                        .passphrase("password")
-                                        .keystore(Resource.create("server.p12")))))
-                .addRouting(GrpcRouting.builder()
-                        .unary(Strings.getDescriptor(),
-                                "StringService",
-                                "Upper",
-                                GrpcTest::upper)
-                        .serverStream(Strings.getDescriptor(),
-                                "StringService",
-                                "Split",
-                                GrpcTest::split)
-                        .clientStream(Strings.getDescriptor(),
-                                "StringService",
-                                "Join",
-                                GrpcTest::join)
-                        .bidi(Strings.getDescriptor(),
-                                "StringService",
-                                "Echo",
-                                GrpcTest::echo));
     }
 
     @Test
